Almost perfect. Just a weird middle artifact near bottom and top middles.
git-svn-id: svn://10.0.50.55/legends/trunk@1249 67207eb2-9daa-044d-9a34-b8567af74898

+++ code/src/GameSatellite/Game_World.h
@@ -320,7 +320,7 @@ public:
 				int UVY = 15-((Index >> 4) & 15);
 				
 				GelColor Col = GEL_RGB_WHITE;
-				if ( (Pos.x.ToInt() == (int)x) && (SubMap.Height()-1-Pos.y.ToInt() == (int)y) ) {
+				if ( (Pos.x.ToInt() == (int)x) && (SubMap.Height()-1-Pos.y.ToInt() == y) ) {
 					Col = GEL_RGB_GREEN;
 				}
 				else {
